server/requirements.txt
dash
dash-bootstrap-components
plotly
pandas
psycopg2-binary